Moved scorecontext.jsx
